+++ InvaderLogicLibrary/IHitBox.cs
@@ -8,10 +8,10 @@ namespace InvaderLogicLibrary
 {
     public interface IHitBox
     {
-        int X { get; set; }
-        int Y { get; set; }
-        int Width { get; set; }
-        int Height { get; set; }
+        double X { get; set; }
+        double Y { get; set; }
+        double Width { get; set; }
+        double Height { get; set; }
 
         bool IsHit(IHitBox hitbox);
     }
